Fix bug, improve security, change stored name
A bug is fixed where on an older version of awk, the record separator
`\0` isn't handled as I intend the code to work. This resulted in
missing newlines when sending the data over curl. The bug is fixed by
changing the record separator.

Security is improved by using `eval` to save variables from the gpg file
instead of using `export`. This way, no child processes will get the
credentials from the environment.

The file saved is changed to `.bb-creds.gpg` to be a more specific
filename.